Implementa sistema completo de roles (admin, bibliotecario, cliente) com authorization por policies e controle de acesso em todas as views e controllers.

+++ atividade01/app/Http/Controllers/AuthorController.php
@@ -11,6 +11,7 @@ class AuthorController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', Author::class);
         $authors = Author::all();
         return view ('authors.index' , compact('authors'));
     }
@@ -20,6 +21,7 @@ class AuthorController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Author::class);
         return view('authors.create');
     }
 
@@ -28,6 +30,7 @@ class AuthorController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Author::class);
         $validateData = $request->validate([
             'name' => 'required|string|max:255',
             'email' => 'required|email|unique:authors,email|max:255',
@@ -45,6 +48,7 @@ class AuthorController extends Controller
      */
     public function show(Author $author)
     {
+        $this->authorize('view', $author);
         return view('authors.show' , compact('author'));
     }
 
@@ -53,6 +57,7 @@ class AuthorController extends Controller
      */
     public function edit(Author $author)
     {
+        $this->authorize('update', $author);
         return view('authors.edit' , compact('author'));
     }
 
@@ -61,6 +66,7 @@ class AuthorController extends Controller
      */
     public function update(Request $request, Author $author)
 {
+    $this->authorize('update', $author);
     $validatedData = $request->validate([
         'name' => 'required|string|max:255',
         'email' => 'required|email|unique:authors,email,' . $author->id . '|max:255',
@@ -78,6 +84,7 @@ class AuthorController extends Controller
      */
     public function destroy(Author $author)
     {
+        $this->authorize('delete', $author);
         $author->delete();
 
         return redirect()->route('authors.index')->with('success', 'Author excluída com sucesso.');
